Refactor package layout; add Settings to menu
