+++ app/src/main/res/layout/fragment_search.xml
@@ -5,7 +5,7 @@
     android:id="@+id/constraintLayout"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.main.PlaceholderFragment">
+    tools:context=".ui.fragment.PlaceholderFragment">
 
     <TextView
         android:id="@+id/section_label"
